Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
textstat
supabase